refactor: rename setting file name
